move shared e2e fixture constants into fixtures.ts
seed.ts goes back to being a pure script with an unconditional top-level
seed() call: with the constants in a side-effect-free module there is
nothing left to import from the seeder, so the realpath entry-point guard
(which silently skipped seeding on a loader/argv quirk) is unnecessary.
