+++ e2e/seed.ts
@@ -3,61 +3,26 @@
  *
  * Seeds the LOCAL test database with realistic fixture data for E2E testing.
  *
- * SAFETY: Database URL is hardcoded to localhost to prevent
- * accidentally running this against a production database (it DELETEs all data).
+ * SAFETY: the database URL (see fixtures.ts) is hardcoded to localhost to
+ * prevent accidentally running this against a production database (it DELETEs
+ * all data).
  *
  * Usage: tsx seed.ts
  */
-import { realpathSync } from "node:fs";
-import { fileURLToPath } from "node:url";
 import pg from "pg";
-
-const DATABASE_URL = "postgres://postgres:postgres@localhost:5432/elmo";
-
-// ---------------------------------------------------------------------------
-// Fixed IDs for test fixtures (so tests can reference them directly)
-// ---------------------------------------------------------------------------
-export const TEST_USER = {
-  email: "e2e@test.local",
-  password: "e2e-test-password-123",
-  name: "E2E Test User",
-} as const;
-
-export const TEST_BRAND_ID = "default";
-export const TEST_BRAND_NAME = "Test Organization";
-export const TEST_BRAND_WEBSITE = "https://example.com";
-
-export const PROMPT_IDS = {
-  branded1: "00000000-0000-0000-0000-000000000001",
-  branded2: "00000000-0000-0000-0000-000000000002",
-  unbranded1: "00000000-0000-0000-0000-000000000003",
-  branded3: "00000000-0000-0000-0000-000000000004",
-  unbranded2: "00000000-0000-0000-0000-000000000005",
-} as const;
-
-export const COMPETITOR_IDS = {
-  competitorA: "00000000-0000-0000-0000-100000000001",
-  competitorB: "00000000-0000-0000-0000-100000000002",
-} as const;
-
-export const REPORT_IDS = {
-  completed: "00000000-0000-0000-0000-300000000001",
-  pending: "00000000-0000-0000-0000-300000000002",
-  processing: "00000000-0000-0000-0000-300000000003",
-  failed: "00000000-0000-0000-0000-300000000004",
-} as const;
-
-// Second tenant — a brand in an org the E2E user is NOT a member of.
-export const NIKE_ORG_ID = "nike";
-export const NIKE_BRAND_ID = "nike";
-export const NIKE_PROMPT_IDS = {
-  training: "00000000-0000-0000-0000-400000000001",
-  lifestyle: "00000000-0000-0000-0000-400000000002",
-} as const;
-export const NIKE_COMPETITOR_IDS = {
-  adidas: "00000000-0000-0000-0000-410000000001",
-  puma: "00000000-0000-0000-0000-410000000002",
-} as const;
+import {
+  COMPETITOR_IDS,
+  DATABASE_URL,
+  NIKE_BRAND_ID,
+  NIKE_COMPETITOR_IDS,
+  NIKE_ORG_ID,
+  NIKE_PROMPT_IDS,
+  PROMPT_IDS,
+  REPORT_IDS,
+  TEST_BRAND_ID,
+  TEST_BRAND_NAME,
+  TEST_BRAND_WEBSITE,
+} from "./fixtures";
 
 // Prompt run IDs (for prompt detail page testing)
 const RUN_IDS = [
@@ -504,26 +469,7 @@ async function seed() {
   }
 }
 
-/**
- * True only when this file is the process entry (`tsx seed.ts`), not when it is
- * imported for its exported fixture constants. Guarding the seed() call on this
- * keeps `import { TEST_* } from "./seed"` side-effect free: otherwise globalSetup
- * and any spec that reads a constant would re-run this destructive seed
- * (DELETE + re-insert) on import and wipe rows out from under running tests.
- */
-function isMainModule(): boolean {
-  const entry = process.argv[1];
-  if (!entry) return false;
-  try {
-    return realpathSync(entry) === realpathSync(fileURLToPath(import.meta.url));
-  } catch {
-    return false;
-  }
-}
-
-if (isMainModule()) {
-  seed().catch((err) => {
-    console.error("Seeding failed:", err);
-    process.exit(1);
-  });
-}
+seed().catch((err) => {
+  console.error("Seeding failed:", err);
+  process.exit(1);
+});
